implement Policies

+++ app/Http/Controllers/UserController.php
@@ -15,35 +15,39 @@ class UserController extends Controller
 {
 	public function destroyUser(Request $request, User $user)
 	{
+		$this->authorize('delete', $user);
+
 		$this->logOut($user, $request);
 
 		User::findOrFail($user->id)->delete();
+
+		event(new DeleteAccount($user));
 
 		return redirect()->route('login.create');
 	}
 
 	public function update(UserRequest $request, User $user)
 	{
+		$this->authorize('update', $user);
+
 		$user->update([
 			'fname' => $request->fname,
 			'lname' => $request->lname,
 			'email' => $request->email
 		]);
 
-		event(new DeleteAccount($user));
-
 		return back()->with(__('app.userSuccessUpdated'));
 	}
 
 	public function updatePassword(updatePassword $request, User $user)
 	{
-		if (!Hash::check($request->oldPassword, $user->password)) {
-			return back()->with("status", "Old Password Doesn't match!");
-		}
+		$this->authorize('update', $user);
 
-		if (Hash::check($request->newPassword, $user->password)) {
+		if (!Hash::check($request->oldPassword, $user->password))
+			return back()->with("status", "Old Password Doesn't match!");
+
+		if (Hash::check($request->newPassword, $user->password))
 			return back()->with("status", "Old Password and new Password are same");
-		}
 
 		User::whereId($user->id)->update([
 			'password' => Hash::make($request->newPassword)
